Improve GPU memory initialization performance
By creating a single big chunk of memory both in GPU and CPU to
allocate the sequences, it's enough with one cudaMemcpy call to transfer
the data.
Additionally, popullating the structs (WF_element array) in host memory with
device pointers, and then do a single cudaMemcpy is much more efficient than
sending the data each time pointers for each elements are generated.